fixed some problems
