+++ src/Components/Input/Input.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+export const Input = ({
+  id,
+  inputClassName,
+  labelClassName,
+  type,
+  name,
+  labelId,
+  content,
+}) => (
+  <>
+    <label htmlFor={labelId} className={labelClassName}>
+      {content}
+    </label>
+    <input id={id} name={name} type={type} className={inputClassName} />
+  </>
+);
